psrsearch

psrsearch is a set of tools for analysing (mostly X-ray) pulsar data and looking for pulsations.

VERY EARLY DRAFT. Mostly for educational purposes only

psrsearch.z_n

psrsearch.z_n_binned

psrsearch.z_n_search

psrsearch.z_n — Method

$Z^2_n$ statistics, à la Buccheri+83, A&A, 128, 245, eq. 2.

Parameters

phase : array of floats The phases of the events n : int, default 2 Number of harmonics, including the fundamental

Returns

z2n : float The Z^2n statistic

Examples

julia> using psrsearch

julia> z_n([10., 0., 0., 0., 0.], 2)
20.0
julia> z_n(ones(10), 2)
40.0
julia> z_n(Array([0.5]), 2)
0.0

psrsearch.z_n_binned — Method

$Z^2_n$ statistic for pulse profiles from binned events

See Bachetti+2021, arXiv:2012.11397

Parameters

profile : array of floats The folded pulse profile (containing the number of photons falling in each pulse bin) n : int Number of harmonics, including the fundamental

Returns

z2_n : float The value of the statistic

Examples

julia> using psrsearch

julia> z_n_binned(zeros(10), 2)
0.0

julia> z_n_binned(ones(10), 2) < 1e-30
true

julia> z_n_binned([10., 0., 0., 0., 0.], 2)
40.0

psrsearch.z_n_search — Method

Calculate the Z^2_n statistics at trial frequencies in photon data.

Parameters

times : array-like the event arrival times n : int the number of harmonics in $Z^2_n$

Other Parameters

fmin : float minimum pulse frequency to search fmax : float maximum pulse frequency to search oversample : float Oversampling factor with respect to the usual 1/T/n rule

Returns

fgrid : array-like frequency grid of the epoch folding periodogram stats : array-like the Z^2_n statistics corresponding to each frequency bin.